+++ verticapy/sql/functions/conditional.py
@@ -59,3 +59,37 @@ str_sql
             i += 2
     expr += " END"
     return str_sql(expr, category)
+
+def decode(expr, *argv):
+    """
+Compares expression to each search value one by one.
+
+Parameters
+----------
+expr: object
+    Expression.
+argv: object
+    Infinite Number of Expressions.
+    The expression generated will look like:
+    even: CASE ... WHEN expr = argv[2 * i] THEN argv[2 * i + 1] ... END
+    odd : CASE ... WHEN expr = argv[2 * i] THEN argv[2 * i + 1] ... ELSE argv[n] END
+
+Returns
+-------
+str_sql
+    SQL expression.
+    """
+    n = len(argv)
+    if n < 2:
+        raise ParameterError(
+            "The number of arguments of the 'decode' function must be greater than 3."
+        )
+    category = to_dtype_category(argv[1])
+    expr = (
+        "DECODE("
+        + str(format_magic(expr))
+        + ", "
+        + ", ".join([str(format_magic(elem)) for elem in argv])
+        + ")"
+    )
+    return str_sql(expr, category)
