+++ verticapy/sql/functions/math.py
@@ -14,16 +14,8 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-
-#
-#
-# Modules
-#
-# VerticaPy Modules
 from verticapy.core.str_sql import str_sql
 from verticapy.sql._utils._format import format_magic
-from verticapy._utils._cast import to_dtype_category
-from verticapy.errors import ParameterError
 
 PI = str_sql("PI()")
 E = str_sql("EXP(1)")
@@ -355,41 +347,6 @@ str_sql
     """
     expr = format_magic(expr)
     return str_sql(f"COT({expr})", "float")
-
-
-def decode(expr, *argv):
-    """
-Compares expression to each search value one by one.
-
-Parameters
-----------
-expr: object
-    Expression.
-argv: object
-    Infinite Number of Expressions.
-    The expression generated will look like:
-    even: CASE ... WHEN expr = argv[2 * i] THEN argv[2 * i + 1] ... END
-    odd : CASE ... WHEN expr = argv[2 * i] THEN argv[2 * i + 1] ... ELSE argv[n] END
-
-Returns
--------
-str_sql
-    SQL expression.
-    """
-    n = len(argv)
-    if n < 2:
-        raise ParameterError(
-            "The number of arguments of the 'decode' function must be greater than 3."
-        )
-    category = to_dtype_category(argv[1])
-    expr = (
-        "DECODE("
-        + str(format_magic(expr))
-        + ", "
-        + ", ".join([str(format_magic(elem)) for elem in argv])
-        + ")"
-    )
-    return str_sql(expr, category)
 
 
 def degrees(expr):
